不能删除root 账号

+++ beike/Repositories/CustomerRepo.php
@@ -66,9 +66,13 @@ class CustomerRepo
     /**
      * @param $id
      * @return void
+     * @throws \Exception
      */
     public static function delete($id)
     {
+        if ($id == 1) {
+            throw new \Exception(trans('admin/customer.cannot_delete_root'));
+        }
         Customer::query()->find($id)->delete();
     }
